Rewrote GLaDOSTeleOp to be command based and removed parts based on current design.
